breaking _type.cc down a bit
